fix: 파이어베이스 데이터 수정 (#44)
