+++ src/api/dataStructure.json
@@ -1,0 +1,54 @@
+{
+  "users": {
+    "<userId>": {
+      "username": "사용자 이름",
+      "email": "이메일",
+      "profileImg": "프로필 이미지 URL",
+      "playlistCount": 0,
+      "totalLikes": 0,
+      "totalForks": 0
+    }
+  },
+  "playlists": {
+    "<playlistId>": {
+      "userId": "<userId>",
+      "title": "플레이리스트 제목",
+      "description": "설명",
+      "category": "카테고리",
+      "createdAt": "타임스탬프",
+      "updatedAt": "타임스탬프",
+      "likeCount": 0,
+      "forkCount": 0,
+      "commentCount": 0,
+      "videoCount": 0,
+      "thumbnailUrl": "썸네일 URL",
+      "isPublic": true,
+      "videos": [
+        {
+          "videoId": "유튜브 비디오 ID",
+          "title": "비디오 제목",
+          "thumbnailUrl": "비디오 썸네일 URL",
+          "duration": 300
+        }
+      ]
+    }
+  },
+  "userPlaylists": {
+    "<userId>": {
+      "created": ["<playlistId1>", "<playlistId2>"],
+      "forked": ["<playlistId5>", "<playlistId6>"],
+      "liked": ["<playlistId3>", "<playlistId4>"]
+    }
+  },
+  "comments": {
+    "<playlistId>": {
+      "<commentId>": {
+        "userId": "<userId>",
+        "username": "사용자 이름",
+        "content": "댓글 내용",
+        "createdAt": "타임스탬프",
+        "updatedAt": "타임스탬프"
+      }
+    }
+  }
+}
