--- src/server/dataStructure.json
@@ -1,52 +1,0 @@
-{
-  "users": {
-    "<userId>": {
-      "username": "사용자 이름",
-      "email": "이메일",
-      "profileImg": "프로필 사진 URL"
-    }
-  },
-  "playlists": {
-    "<userId>": {
-      "<playlistId>": {
-        "title": "플레이리스트 제목",
-        "createdAt": "timestamp",
-        "likes": 0,
-        "commentCount": 0,
-        "thumbnailUrl": "플레이리스트 썸네일 URL",
-        "videos": {
-          "<videoId>": {
-            "addedAt": "timestamp",
-            "order": 1
-          }
-        }
-      }
-    }
-  },
-  "videos": {
-    "<videoId>": {
-      "title": "비디오 제목",
-      "thumbnailUrl": "비디오 썸네일 URL",
-      "uploader": "업로더 이름"
-    }
-  },
-  "forkedPlaylists": {
-    "<userId>": {
-      "<playlistId>": true
-    }
-  },
-  "likedPlaylists": {
-    "<userId>": {
-      "<playlistId>": true
-    }
-  },
-  "comments": {
-    "<playlistId>": {
-      "<commentId>": {
-        "userId": "<userId>",
-        "content": "댓글 내용",
-        "createdAt": "timestamp"
-      }
-    }
-  }
-}
