--- src/server/firestoreService.ts
@@ -1,14 +1,0 @@
-import { collection, getDocs, DocumentData } from 'firebase/firestore';
-
-import { db } from './firebase';
-
-export const fetchCollection = async (collectionName: string): Promise<DocumentData[]> => {
-  try {
-    const querySnapshot = await getDocs(collection(db, collectionName));
-    const data = querySnapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
-    return data;
-  } catch (error) {
-    console.error(`Error fetching collection ${collectionName}:`, error);
-    throw error;
-  }
-};
